requirements.txt
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
aiosqlite>=0.20
pydantic>=2.7
pydantic-settings>=2.2
python-multipart>=0.0.9
Pillow>=10.3
